backend/pyproject.toml
[tool.poetry]
name = "respiratory-risk-backend"
version = "0.1.0"
description = "Backend service for respiratory risk API"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.100.0"
uvicorn = { extras = ["standard"], version = "^0.23.0" }
httpx = "^0.25.0"
pydantic = "^2.2.0"
pydantic-settings = "^2.0.0"
aiocache = { extras = ["redis"], version = "^0.11.1" }
aioredis = "1.3.1"
python-dotenv = "^1.0.0"
asyncpg = "^0.28.0"
sqlalchemy = "^2.0"
psycopg2-binary = "^2.9"

# Observability libraries
opentelemetry-api = "^1.23.0"
opentelemetry-sdk = "^1.23.0"
opentelemetry-instrumentation-fastapi = "^0.44b0"
opentelemetry-instrumentation-httpx = "^0.44b0"
prometheus-fastapi-instrumentator = "^6.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-asyncio = "^0.20.0"
requests = "^2.32.4"
# Static analysis and formatting tools
black = "^24.4.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py312"]
